Add singly-linked lists
As there are some hold-outs who don't want the RAM overhead of ns_list's
deluxe implementation, fork ns_list to two new singly-linked
implementations.

ns_slist has singly-linked nodes, but retains the dual head and tail
pointers, making it equivalent to BSD's STAILQ or SIMPLEQ.

ns_hslist has singly-linked nodes, and only a single head pointer,
making it equivalent to BSD's SLIST.

Naming is derived from Linux - its `list` is equivalent to our
fully-featured list (albeit implemented as circular). It has a `hlist`
that is doubly-linked but with only a head pointer, designed for hash
tables.

So our variants become `slist` for the single-linking and `hslist` for
head-only/hash-suitable and singly-linked. The naming leaves room for
the Linux-style `hlist` which is not currently provided.